dashboard/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
supabase==2.4.0
python-dotenv==1.0.0
httpx==0.27.0
